Create @backend
Still working on it...There is a need for a couple of features and changes to be made not yet been done. As well as the variable naming linked from the SQL code. Thanks so much for the team effort.
Though the front-end features and a Server are needed urgently so that it can run through from any Computer and location for all to see.